Dimension Reduction and Estimation Methods

Rdimtools is an R suite of a number of dimension reduction and estimation methods implemented using RcppArmadillo for efficient computations. Please see the section below for the complete composition of this package and what we can provide in a unifying interface across many methods.

Composition of the package

The package consists of following families of functions whose names start with do., est., and aux. for performing dimension reduction/manifold learning, estimating intrinsic dimension, and some efficient implementations of other useful methods respectively.

(1) do. family for dimension reduction algorithms

do. functions are for dimension reduction (or, manifold learning) methods. A simple taxonomy of the methods would be to categorize based on the linearity of embedding mappings. In the table below, TYPE represents whether it is supervised (S), semisupervised (SS), or unsupervised (U).

For linear methods, we have

FUNCTION	TYPE	ALGORITHM

do.adr	U	Adaptive Dimension Reduction

do.ammc	S	Adaptive Maximum Margin Criterion

do.anmm	S	Average Neighborhood Margin Maximization

do.asi	U	Adaptive Subspace Iteration

do.bpca	U	Bayesian Principal Component Analysis

do.cca	S	Canonical Correlation Analysis

do.cnpe	U	Complete Neighborhood Preserving Embedding

do.crp	U	Collaborative Representation-based Projection

do.cscore	S	Constraint Score

do.cscoreg	S	Constraint Score using Spectral Graph

do.dagdne	S	Double-Adjacency Graphs-based Discriminant Neighborhood Embedding

do.dne	S	Discriminant Neighborhood Embedding

do.disr	U	Diversity-Induced Self-Representation

do.dspp	S	Discriminative Sparsity Preserving Projection

do.elde	S	Exponential Local Discriminant Embedding

do.elpp2	U	Enhanced Locality Preserving Projection (2013)

do.enet	S	Elastic Net Regularization

do.eslpp	S	Extended Supervised Locality Preserving Projection

do.extlpp	U	Extended Locality Preserving Projection

do.fa	U	(Exploratory) Factor Analysis

do.fscore	S	Fisher Score

do.fssem	U	Feature Subset Selection using Expectation-Maximization

do.ica	U	Independent Component Analysis

do.isoproj	U	Isometric Projection

do.kmvp	S	Kernel-Weighted Maximum Variance Projection

do.kudp	U	Kernel-Weighted Unsupervised Discriminant Projection

do.lasso	S	Least Absolute Shrinkage and Selection Operator

do.lda	S	Linear Discriminant Analysis

do.ldakm	U	Combination of LDA and K-means

do.lde	S	Local Discriminant Embedding

do.ldp	S	Locally Discriminating Projection

do.lea	U	Locally Linear Embedded Eigenspace Analysis

do.lfda	S	Local Fisher Discriminant Analysis

do.llp	U	Local Learning Projections

do.lltsa	U	Linear Local Tangent Space Alignment

do.lmds	U	Landmark Multidimensional Scaling

do.lpca	U	Locally Principal Component Analysis

do.lpe	U	Locality Pursuit Embedding

do.lpfda	S	Locality Preserving Fisher Discriminant Analysis

do.lpmip	U	Locality-Preserved Maximum Information Projection

do.lpp	U	Locality Preserving Projection

do.lqmi	S	Linear Quadratic Mutual Information

do.lscore	U	Laplacian Score

do.lsda	S	Locality Sensitive Discriminant Analysis

do.lsdf	SS	Locality Sensitive Discriminant Feature

do.lsir	S	Localized Sliced Inverse Regression

do.lspe	U	Locality and Similarity Preserving Embedding

do.lspp	S	Local Similarity Preserving Projection

do.mcfs	U	Multi-Cluster Feature Selection

do.mds	U	(Metric) Multidimensional Scaling

do.mfa	S	Marginal Fisher Analysis

do.mlie	S	Maximal Local Interclass Embedding

do.mmc	S	Maximum Margin Criterion

do.mmp	SS	Maximum Margin Projection

do.mmsd	S	Multiple Maximum Scatter Difference

do.modp	S	Modified Orthogonal Discriminant Projection

do.msd	S	Maximum Scatter Difference

do.mvp	S	Maximum Variance Projection

do.nolpp	U	Nonnegative Orthogonal Locality Preserving Projection

do.nonpp	U	Nonnegative Orthogonal Neighborhood Preserving Projections

do.npca	U	Nonnegative Principal Component Analysis

do.npe	U	Neighborhood Preserving Embedding

do.nrsr	U	Non-convex Regularized Self-Representation

do.odp	S	Orthogonal Discriminant Projection

do.olda	S	Orthogonal Linear Discriminant Analysis

do.olpp	U	Orthogonal Locality Preserving Projection

do.onpp	U	Orthogonal Neighborhood Preserving Projections

do.opls	S	Orthogonal Partial Least Squares

do.pca	U	Principal Component Analysis

do.pflpp	U	Parameter-Free Locality Preserving Projection

do.pls	S	Partial Least Squares

do.ppca	U	Probabilistic Principal Component Analysis

do.rlda	S	Regularized Linear Discriminant Analysis

do.rndproj	U	Random Projection

do.rpcag	U	Robust Principal Component Analysis via Geometric Median

do.rsir	S	Regularized Sliced Inverse Regression

do.rsr	U	Regularized Self-Representation

do.sammc	SS	Semi-Supervised Adaptive Maximum Margin Criterion

do.save	S	Sliced Average Variance Estimation

do.sda	SS	Semi-Supervised Discriminant Analysis

do.sdlpp	U	Sample-Dependent Locality Preserving Projection

do.sir	S	Sliced Inverse Regression

do.slpe	S	Supervised Locality Pursuit Embedding

do.slpp	S	Supervised Locality Preserving Projection

do.spc	S	Supervised Principal Component Analysis

do.spca	U	Sparse Principal Component Analysis

do.specs	S	Supervised Spectral Feature Selection

do.specu	U	Unsupervised Spectral Feature Selection

do.spp	U	Sparsity Preserving Projection

do.spufs	U	Structure Preserving Unsupervised Feature Selection

do.ssldp	SS	Semi-Supervised Locally Discriminant Projection

do.udfs	U	Unsupervised Discriminative Features Selection

do.udp	U	Unsupervised Discriminant Projection

do.ulda	S	Uncorrelated Linear Discriminant Analysis

Also, we have nonlinear methods implemented

FUNCTION	TYPE	ALGORITHM

do.bmds	U	Bayesian Multidimensional Scaling

do.cge	SS	Constrained Graph Embedding

do.cisomap	U	Conformal Isometric Feature Mapping

do.crca	U	Curvilinear Component Analysis

do.crda	U	Curvilinear Distance Analysis

do.dm	U	Diffusion Maps

do.dve	U	Distinguishing Variance Embedding

do.fastmap	U	FastMap

do.idmap	U	Interactive Document Map

do.iltsa	U	Improved Local Tangent Space Alignment

do.isomap	U	Isometric Feature Mapping

do.ispe	U	Isometric Stochastic Proximity Embedding

do.keca	U	Kernel Entropy Component Analysis

do.klde	S	Kernel Local Discriminant Embedding

do.klfda	S	Kernel Local Fisher Discriminant Analysis

do.klsda	S	Kernel Locality Sensitive Discriminant Analysis

do.kmfa	S	Kernel Marginal Fisher Analysis

do.kmmc	S	Kernel Maximium Margin Criterion

do.kpca	U	Kernel Principal Component Analysis

do.kqmi	S	Kernel Quadratic Mutual Information

do.ksda	SS	Kernel Semi-Supervised Discriminant Analysis

do.lamp	U	Local Affine Multidimensional Scaling

do.lapeig	U	Laplacian Eigenmaps

do.lisomap	U	Landmark Isometric Feature Mapping

do.lle	U	Locally Linear Embedding

do.llle	U	Local Linear Laplacian Eigenmaps

do.ltsa	U	Local Tangent Space Alignment

do.mve	U	Minimum Volume Embedding

do.mvu	U	Maximum Variance Unfolding / Semidefinite Embedding

do.nnp	U	Nearest Neighbor Projection

do.plp	U	Piecewise Laplacian Projection

do.ree	U	Robust Euclidean Embedding

do.rpca	U	Robust Principal Component Analysis

do.sammon	U	Sammon Mapping

do.sne	U	Stochastic Neighbor Embedding

do.spe	U	Stochastic Proximity Embedding

do.splapeig	S	Supervised Laplacian Eigenmaps

do.spmds	U	Spectral Multidimensional Scaling

do.tsne	U	t-distributed Stochastic Neighbor Embedding

(2) est. family for intrinsic dimension estimation algorithms

est. family of functions include,

FUNCTION	TYPE	ALGORITHM

est.boxcount	G	Box-Counting Dimension

est.clustering	G	Clustering-based Estimation

est.correlation	G	Correlation Dimension

est.danco	G	Dimensionality from Angle and Norm Concentration

est.made	G/P	Manifold-Adaptive Dimension Estimation

est.gdistnn	G/P	Graph Distance based on Manifold Assummption

est.mindkl	G	Minimum Neighbor Distance with Kullback Leibler Divergence

est.mindml	G	Minimum Neighbor Distance with Maximum Likelihood

est.mle1	G	MLE using Poisson Process

est.mle2	G	MLE using Poisson Process with Bias Correction

est.nearneighbor1	G	Near-Neighbor Information

est.nearneighbor2	G	Near-Neighbor Information with Bias Correction

est.incisingball	G	Estimation using Incising Ball

est.packing	G	Estimation using Packing Numbers

est.pcathr	G	PCA Thresholding with Accumulated Variance

est.twonn	G	Minimal Neighborhood Information

est.Ustat	G	Convergence Rate of U-statistic

where the taxonomy is of global(G) or pointwise(P). Global methods return a single estimated dimension of the data manifold, whereas Pointwise methods return locally estimated intrinsic dimension at each point.

(3) oos. family for out-of-sample predictions

If the original dimension reduction method was linear-type, then you could use oos.linear function.

FUNCTION	ALGORITHM

oos.linear	Projection using pre-computed mappings.

Regardless of the types of previous manifold learning methods, belows are general out-of-sample methods.

FUNCTION	ALGORITHM

oos.linproj	Linear Projection.

(4) aux. functions

Some auxiliary functions (aux.) are also provided,

FUNCTION	DESCRIPTION

aux.gensamples	generates samples from predefined shapes

aux.graphnbd	builds a neighborhood graph given certain criteria

aux.kernelcov	computes a centered gram matrix with 20 kernels supported

aux.preprocess	performs preprocessing of centering, decorrelating, or whitening

aux.shortestpath	Floyd-Warshall algorithm (it's Fast!)

aux.pkgstat	reports the number of functions available for each category