MINDml

It is a minimum neighbor distance estimator of the intrinsic dimension based on Maximum Likelihood principle.

est.mindml(X, k = 5)

Arguments

X
an \((n\times p)\) matrix or data frame whose rows are observations.

k
the neighborhood size for defining locality.

Value

a named list containing containing

estdim

the global estimated dimension.

References

Lombardi G, Rozza A, Ceruti C, Casiraghi E, Campadelli P (2011). “Minimum Neighbor Distance Estimators of Intrinsic Dimension.” In Gunopulos D, Hofmann T, Malerba D, Vazirgiannis M (eds.), Machine Learning and Knowledge Discovery in Databases, volume 6912, 374--389. Springer Berlin Heidelberg, Berlin, Heidelberg. ISBN 978-3-642-23782-9 978-3-642-23783-6, doi: 10.1007/978-3-642-23783-6_24 .

See also

Examples

# \donttest{
## create 3 datasets of intrinsic dimension 2.
X1 = aux.gensamples(dname="swiss")
X2 = aux.gensamples(dname="ribbon")
X3 = aux.gensamples(dname="saddle")

## acquire an estimate for intrinsic dimension
out1 = est.mindml(X1, k=10)
out2 = est.mindml(X2, k=10)
out3 = est.mindml(X3, k=10)

## print the results
sprintf("* est.mindml : estimated dimension for 'swiss'  data is %.2f.",out1$estdim)

#> [1] "* est.mindml : estimated dimension for 'swiss'  data is 2.00."

sprintf("* est.mindml : estimated dimension for 'ribbon' data is %.2f.",out2$estdim)

#> [1] "* est.mindml : estimated dimension for 'ribbon' data is 2.06."

sprintf("* est.mindml : estimated dimension for 'saddle' data is %.2f.",out3$estdim)

#> [1] "* est.mindml : estimated dimension for 'saddle' data is 2.15."

# }